SCUMM: Fix undefined va_arg() behavior
The second parameter to va_arg() needs to be "compatible with the type
of the actual next argument (as promoted according to the default
argument promotions)" or "random errors will occur".

Apparently what that means is that you can't use any integer types
smaller than a signed or unsigned int. In my case, ScummVM crashed with
an illegal instruction error.

+++ engines/scumm/players/player_mac_indy3.cpp
@@ -840,10 +840,10 @@ void I3MFourToneSynthDriver::setParameter(ParaType type, ...)  {
 
 	switch (type) {
 	case kDuration:
-		setDuration(va_arg(arg, uint16));
+		setDuration((uint16)va_arg(arg, uint));
 		break;
 	case kChanRate:
-		setRate(chan, va_arg(arg, uint16));
+		setRate(chan, (uint16)va_arg(arg, uint));
 		break;
 	case kChanWaveform:
 		setWaveForm(chan, ptr, va_arg(arg, uint32));
@@ -901,9 +901,9 @@ void I3MSquareWaveSynthDriver::setParameter(ParaType type, ...)  {
 	
 	switch (type) {
 	case kSwTriplet:
-		a = va_arg(arg, uint16);
-		b = va_arg(arg, uint16);
-		pushTriplet(a, b, va_arg(arg, uint16));
+		a = (uint16)va_arg(arg, uint);
+		b = (uint16)va_arg(arg, uint);
+		pushTriplet(a, b, (uint16)va_arg(arg, uint));
 		break;
 	default:
 		break;
